+++ api/api/Domain/Entities/Song.cs
@@ -15,7 +15,7 @@ public class Song
     public required long GenreId { get; set; }
     public Genre Genre { get; set; }
     public required TimeSpan Duration { get; set; }
-    [Range(1, 4)] public required decimal MediaRating { get; set; }
+    [Range(1, 4)] public required int MediaRating { get; set; }
     [Range(0, long.MaxValue)] public required long Streams { get; set; }
     public required DateTime AddedAt { get; set; } = DateTime.Now;
     public ICollection<Rating> Ratings { get; set; }
